Make `deploy` implicit.
There was a mismatch between the recently introduced desired stare
pipeline management architecture and the external API.  Internally,
there are only three possible desired state (shutdown, running, and
paused), but the API still distinguished the `/deploy` action,
which validated and committed pipeline configuration before initializing
the pipeline and the `/pause` action which also set the desired state to
`paused` but was only available in the `running` state (and even that
was not properly enforced).

We clean things up by establishing 1-1 correspondence between actions
and states.  In particular, `pause` and `start` actions are available in
all non-temporary states and do whatever is needed to take the pipeline
to the `paused` or `running` state respectively, including validating
and committing the pipeline when necessary.  We still keep the `/deploy`
action in the API to avoid breaking the UI, but it is now a synonym
of `/pause` and can be safely removed once no longer used.

This will allow simplifying the UI, which can now simply issue the
`/start` command to the pipeline instead of `/deploy` followed by
`/start`.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>