Propagate pipeline panics to the user.
A panic in a DBSP worker thread surfaced as a log message, leaving the
pipeline in a broken state, but did not propagate to the manager, which
still saw the pipeline as healthy (since its HTTP server remains
operational in this case).

This commit extends the server logic in the `adapters` crate to handle
fatal errors by setting pipeline status to Failed.  In this state the
pipeline returns an error in response to all endpoint invocations.
As soon as the manager polls the pipeline (currently every 10s), it
will change its `current_status` to `Failed` in the DB, which will be
reported to the user whenever they query the state of the pipeline.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>