Distinguish empty batch from keep-alive message.
Keep-alive messages are generated every few seconds when the
output stream doesn't have any useful payload as a workaround
for an actix bug.  When this message is generated in response
to a quantile or neighborhood query before the actual query
result, this causes the UI to temporarily show the table as
being empty.

To help the UI distinguish an empty chunk from a keep-alive
message, we omit the payload field in the latter.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>