Eliminate dependency on arcstr patch that isn't upstream yet.
DBSP relies on bincode version 2.0.0-rc3, which was released at the end of
March 2023.  DBSP also uses arcstr and we want to serialize and
deserialize `ArcStr`s with bincode.  Unfortunately, arcstr doesn't support
bincode 2.0 yet and won't until bincode 2.0 finally gets released[1], which
could be a long time, so we have a patch in Cargo.toml to support that.

We'd like to publish DBSP to crates.io[2], but we can't if it depends on
patches, so it seems best to avoid them.  This commit drops the patch and
switches to using #[bincode(with_serde)] to serialize ArcStr fields.

[1]: https://github.com/thomcc/arcstr/pull/45
[2]: https://github.com/feldera/dbsp/issues/14

Signed-off-by: Ben Pfaff <blp@feldera.com>
Fixes: https://github.com/feldera/dbsp/issues/395