dbsp_adapters: Copy in `Deserializer` to enable using upstream `csv` crate.
The `dbsp_adapters` crate needs a `Deserializer` for CSV data.  The `csv`
crate has such a `Deserializer` internally that it does not make public.
However, its `Deserializer` implementation only uses public APIs, so
we can just copy it into our own tree.  That's what this does.

This allows us to drop the use of our forked `csv` crate, and this commit
also does that.

Signed-off-by: Ben Pfaff <blp@feldera.com>
Fixes: https://github.com/feldera/dbsp/issues/394
Signed-off-by: Ben Pfaff <blp@feldera.com>