Compiler: support deserializing into a struct.
This commit prepares the SQL compiler for deserializing JSON data.

The compiler internally represents database records as tuples, which
works fine, except that it is impossible to deserialize them directly
from the most natural JSON representation as a map:

```json
{
   "column1": ...,
   "column2": ...
}
```

This commit addresses this by additionally generating a `struct TABLE`
type for each table and view and a `From` impl of the form
`impl From<TABLE> for TupleNN<...>`.  With this we can deserialize a
JSON map into `struct TABLE` and then convert it into the tuple
representation.

Signed-off-by: Mihai Budiu <mbudiu@feldera.com>