pipeline-manager: have compiler reconcile local state with API state at startup
The compiler service now attempts to reconcile its local state with the program objects
available in the DB. If it sees any evidence of programs without local
versioned binaries, but have started or completed compilation, the service re-queues
the program for compilation.

The PR also adds some tests for the compiler service to check this behavior.

Addresses the compiler issue in #235

Signed-off-by: Lalith Suresh <lalith@feldera.com>