adapters: remove the "server" feature.
This feature gate is meant to make the HTTP server part of the `adapters`
crate optional.  At this point I don't see us using the crate without
the HTTP functionality, which percolates more and more throughout the
various APIs in the crate, and it is a pain to continue feature gating all
server-related code.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>