Improve handling and reporting of parse errors.
The parser API was originally designed assuming that the parser gives up
parsing the current buffer upon hitting the first parse error.  This
led to unpredictable behavior where, on error, the set of records that
were successfully ingested depended on how the input data stream got
broken up into fragments (which is non-deterministic for, e.g., HTTP,
file, and URL transports).

We change this behavior so that the parser continues parsing after
hitting an error and reports all errors it has encountered at the end.
We also introduced a strongly typed representation for parser errors,
which includes event number relative to the start of the stream.
This info can be used by the client to identify records that have been
successfully ingested.

We also extend the JSON parser to ingest arrays of events atomically.
The JSON parser can be configured to ingest events grouped into JSON arrays.
With this commit, we change its behavior to either ingest all records
in the array or none if a parser error occurs while processing the
array.  We may want to make this behavior configurable in the future
or add other atomic delivery mechanism (e.g., some flavor of
transactions), but that's beyond the scope of this commit.

TODOs:
- The next step in improving parser error reporting is to push these
  errors to a DBSP stream, so that they can be stored and browsed as
  a regular table.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>