dbsp_adapters: Better support pause and resume in `url` adapter.
Before this commit, the `url` adapter connected when it was created and
disconnected only when destroyed.  If the adapter was paused in the middle,
the server might disconnect due to idle.

With this commit, the `url` adapter connects when unpaused and disconnects
when paused.  When the adapter is unpaused again, it reconnects and skips
past any data read before.  This should avoid idle disconnects.

Signed-off-by: Ben Pfaff <blp@feldera.com>