dbsp: Better caching of `trace`, `integrate_trace`.
Improve circuit caching so that `trace` and `integrate_trace` operators
can reuse each other's results when possible.

The `trace` operator computes the trace of the input stream across all
observed timestamps.  `integrate_trace` computes the integral of the input
stream during a single run of the innermost circuit.  These happen to be
identical in the context of the root circuit and so the output of one of
these two operators can be reused instead of the other.  This didn't
previously work because the two operators used different cache tags
(`TraceId` and `IntegrateTraceId`).  We unify the two and rely on result
types (which are recorded as part of the cache) to distinguish them
when they are not identical (e.g., in a nested circuit).

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>