integration-test: fix local pipeline manager initialization regression
We removed the daemonization feature in #447. As part of that change, we
reworked a few instances where we'd explicitly create an async runtime after
daemonization could happen. We missed to do so in
`initialize_local_dbsp_instance()`, which broke that function. It was used by
the integration tests to bring up a local pipeline manager for convenience; it
panicked with the "cannot start a runtime from within a runtime" error.

Fixing this was not as simple as just removing the extra runtime being created.
This is because we cannot simply reuse the tokio runtime instance created by
the test (e.g., the one implicitly created via [actix_web::test]) to create the
compiler, local runner and api futures.  The reason is that when the first
test completes, these futures will get cancelled, leading to some confusing
behavior when running integration tests.

To avoid that problem, and for general integration test hygiene, we force
another tokio runtime creation to run the required server processes. We
obviously can't create one runtime within another, so the easiest way to work
around that is to do so from within an invocation of std::thread::spawn().

Signed-off-by: Lalith Suresh <lalith@feldera.com>