Use upsert semantics for input tables.
We change the type of input handles from `ZSetHandle` to `SetHandle`,
which uses upsert semantics for input events, i.e., double insertions
get ignored.  Likewise, and attempt to delete a non-existent or
previously deleted entry gets ignored.  The main reason for this change
is that it prevents clients from creating entries with negative weights,
which work in theory, but are never created intentionally in any
scenario I can think of and can producer surprising outputs.

Although upsert operators maintain the integral of the input trace, this
change does not increase the memory footprint of the pipeline, since we
already build this integral for debugging purposes (as part of
`neighborhood` and `quantiles` operators) and all these operators should
reuse the same instance of the integral.

Several future TODOs:

- Use `MapHandle`'s for input tables with primary keys.  These also have
  upsert semantics but support deletion by key and allow changing the
  value associated with a key with a single event.

- Insert-only collections (event streams) don't require upserts.

- We currently drop noop commands (double insertions and deletions)
  silently.  We may want to provide an option to log such events somehow
  (e.g., by writing them to a separate stream).

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>