dbsp: Switch from `bincode` to `rkyv` and require it for all data.
`bincode` is difficult to apply to types from crates that we don't
control, because of the orphan rule for Rust trait implementations.
This is made worse by `bincode` v2 being in perpetual pre-release, so
crate owners aren't willing to implement its traits for their types.

`rkyv`, on the other hand, is not in pre-release, and crate owners are
more willing to implement its traits.  It also is faster, although
that's counterbalanced by its demand for aligned data which RPC and
database implementations don't like to provide (so you end up doing data
copies).

This switches from `bincode` to `rkyv` across the tree.

Fixes: https://github.com/feldera/dbsp/issues/420

Signed-off-by: Ben Pfaff <blp@feldera.com>