Fixup demo script to run manager in foreground.
The pipeline manager no longer runs as a background daemon process.
This broke the `demo.sh` script, which used to wait for the manager
process to return control to the script before instantiating the demos.

We change the demo flow to run the manager in the foreground and spawn
processes to crate demo pipelines in the background.

We also change `Makefile.toml` to run the manager in the background.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>