dbsp: `update_set` operator.
Added a specialized form of `upsert` operator that works on sets (i.e.,
`Val=()`).  We previously used the same implementation in both cases,
which always represents the trace as `Spine<OrdIndexedZSet>`, which
is less efficient that `Spine<OrdZSet>` for sets. More importantly,
this adversely effects caching, since most operators that work with sets
expect `OrdZSet`.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>