Avoid no-op distinct operators.
`distinct` is a no-op when it is applied to the output of `aggregate`,
`rolling_aggregate`, `upsert` or any other operator that outputs
collections with unit weights only.  To avoid the compute and memory
cost of `distinct`, we label such streams as the distinct versions of
themselves and skip instantiating the distinct operator for them.

Addresses #45.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>