Use insert/delete format for JSON outputs.
There are several issues with the weighted update format:

- It can be tricky for users to understand.
- It doesn't match the default input format.
- Most downstream consumers (e.g., a conventional database)
  don't naturally support this format.

We switch to using the insert/delete format and simply replicate each
record N times, where N is the weight of the record.

One caveat is that weights can be very large.  We don't want to output
for instance a trillion identical records, so we bound the largest
supported weight (currently the bound is 1M and is not configurable)
and return an error if the bound is exceeded.

In the future we will make the output format configurable, similar to
how the input format is already configurable (at the very least we will
need to support various CDC formats).

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>