Pause SecOps generator when demo is not running.
The SecOps generator used to run continuously whether the demo was
running or not.  This created a small but non-trivial I/O and CPU load
on the system.

The fix is to only run the generator when the pipeline is running and is
close to exhausting all previously generated messages.  We achieve this
but monitoring the consumer group offset.  When the offset is more
than 1000 messages behind the producer, the generator stops.

This requires the producer to know consumer group id, so we use a fixed
group id for one of the topics in the demo.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>