adapters: Be more generous about initial timeout in url adapter pause test.
The "pause" test checks that pause and resume of HTTP download in the url
adapter works successfully.  It's inherently a timing based test, and CI
systems are usually busy, so it's necessary to define some timeouts.  I
thought that the initial timeout I set was generous, but it keeps failing
in CI.  On the assumption that this is due to busy CI, and not to a bug in
the url adapter, this commit increases the timeout further.

Signed-off-by: Ben Pfaff <blp@feldera.com>