Reduce log level for some errors.
Some of the things we report as HTTP errors are benign events that are
part of the normal operation.  One example is retrieving a program by
name to check if it exists.  Another one is probing pipeline status to
determine whether it's finished initializing.

Logging such errors at the ERROR log level is confusing to the user.
This commit makes error log level configurable by extending the
`DetaileError` trait, which all our errors implment, with the
`log_level()` method.  It return `Level::Error`, by default, but can
be customized.  For now I reduced the log level of some obvious
candidates to Info.  Going forward, we now have a way to fine tune this
behavior.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>