dbsp: Get the `gdelt` benchmark compiling again.
This implements the `rkyv` traits for the data types that `gdelt`
implements, starts using the DBSP newtype for `ArcStr`, and
switches from `hashbrown` to standard collections.  The latter is
necessary because `hashbrown` only supports `rkyv` as of its
latest release (0.14.0) but we have dependencies that pin `rkyv`
to earlier versions.

Signed-off-by: Ben Pfaff <blp@feldera.com>