dbsp: Remove custom join implementation from `gdelt` benchmark.
This removes hundreds of lines of source code and it does not have much of
an effect on runtime in my testing.

Total time before this commit: 137.565ms
Total time after this commit: 137.323ms

I also went back to before the distribution code was committed and
fixed the build for this benchmark there and ran it.  There was a
lot of variance from run to run but I saw about 135 seconds runtime
as average.

Signed-off-by: Ben Pfaff <blp@feldera.com>