dbsp-adapters: Add trivial Kafka input test so we can fail fast.
When there's some kind of problem with redpanda during a CI build,
the proptest_kafka_input failure spews 80,000 lines of garbage.  But the
test fails with a much simpler input just the same way, but without all
the spew.  This commit adds such a test to make CI failures easier to
debug.

This looks like a big change but really most of it is just moving code
around and re-indenting.  I also had to give the two tests different Kafka
topic names so they don't interfere with one another.

Signed-off-by: Ben Pfaff <blp@feldera.com>