Enforce max Kafka message size.
Some transports, e.g., Kafka, can impose a limit on the number of bytes
in a message.  In case of Kafka, the limit is specified via the
`message.max.bytes` config option.  The encoder must respect this limit,
otherwise a large output batch can generate messages that are too large.
However, until now we did not have a mechanism to communicate the limit
to the encoded.

We add a new `max_buffer_size_bytes()` method to `trait OutputConsumer`
and `trait OutputEndpoint`.  We also modify the CSV and JSON encoders
(the only two formats we currently support) to respect the message size
limit.

Finally, we extend the Kafka test harness to test different format and
message size limits.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>