Kafka tests: wait for new topics to be available.
Just because `create_topics` succeeded doesn't guarantee the new topics
are avilable (see https://github.com/confluentinc/confluent-kafka-python/issues/524).

This caused failures in Kafka tests when running on CI machines with
slow disks.  The test harness deletes an existing instance of the topic
before creating a new one, which seems to cause the delay because the
previous instance is still being deleted when the
`create_topics` command arrives.

I added a probing loop that checks whether the topic is available to
consume from for up to 20seconds to the test harness.

Question is, do we want to do something like this in the Kafka adapter.
I don't want to do this by default, so that the adapters fails quickly
if the topic truly doesn't exist (common case).  But this could be a
config option in the future.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>