Kafka: Allow overriding auto-commit settings.
Commit 0b7edc2fbbb6a328679af941fb981dc5b563b15e added a workaround for a
glitch in rdkafka, which required disabling the auto-commit behavior.
Unfortunately, this broke the secops demo simulator, which monitors
consumer offset in order to determine when to resume sending data to the
topic.  If the consumer doesn't ever commit its offset, the simulator
never resumes after writing some initial data.

There might be other use cases where the user wants the pipeline to commit
its offset.  This should be safe in most cases, the rdkafka bug only
affects the situation where the broker or the topic disappears right
before the pipeline shuts down.

This commit keeps the default behavior of disabling auto-commit, but
allows the user to override it using connector config.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>