dbsp_adapters: Check even more carefully that Kafka topics are created.
I found that even the previous check wasn't enough.  This fixes the
problem for me.

This solves a problem for upcoming code that does more with topics and
partitions immediately after Kafka comes up.  I have not seen this problem
with the existing code.

This technique doesn't need to create a consumer at all, so this commit
removes the code that does that.

I changed the backoff from 1 second to 100 ms plus doubling after each
failure.  This is because I consistently see at least one backoff in my
testing and 100 ms was consistently enough, so this avoids slowing the
tests more than necessary.

Signed-off-by: Ben Pfaff <blp@feldera.com>