adapters: Fix the number of actix threads.
By default actix starts the number of threads equal to the number of cores,
which is an overkill and can lead to file descriptor exhaustion when running
multiple pipelines on systems with >100 cores.  We use a fixed number of
threads instead.  Four should be more than enough, but we can make it
configurable if needed.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>