Debezium MySQL demo.
This commit adds a new demo program that ingests data into Feldera
from a MySQL database using Debezium.

We configure the following services:

- Redpanda
- Kafka Connect with Debezium connector using the RedPanda service as
  its internal state store.
- MySQL container with a simple database from the Debezium MySQL tutorial

and create a Debezium connector that streams updates to the database to
RedPanda topics (one topic per table).  All of this setup is done outside
of Feldera.  Inside Feldera, we create a SQL program that mirrors the MySQL
program and build a pipeline with one Kafka connector per topic.
Connectors are configured to parse the Debezium JSON format (I
implemented support for it earlier and only had to fix it up slightly).

With this setup we are able to ingest the contents of all tables except the
one table that has a field of type `DATE` (see below).

I added a minimal test of the above config to CI.  It only starts the
demo, but doesn't look at the outputs or do anything interesting with
it.  I didn't want to spend too much time on this until we have an
end-to-end architecture.

TODOs:

- Parsing Debezium-generated JSON.  The format is generally
  straightforward and we already mostly support it.  However Debezium
  has a very unfortunate way to encode temporal types.  Not only they
  use numbers instead of strings (that wouldn't be too bad), but the
  semantics of the number depends on (1) the exact column type, (2) connector
  config, and (3) source database (i.e., MySQL, Postgres, etc.).  The encoding
  of decimals is also tricky.

- Replace the Debezium tutorial container with our own MySQL container
  configured the way we expect Feldera users to configure their MySQL
  instances for Debezium.

- UI for creating Feldera Debezium connectors (as explained above, it's
  just a special case of a Kafka + JSON connector).

- We probably want to look into Avro support.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>