A framework for configurable deserialization.
In Feldera, we deal with different encodings of the same datatype, e.g.,
timestamps can be represented as strings formatted using different standards
or as numbers.  In the latter case, these numbers can represent times in
microseconds or milliseconds.  Different data sources can define their
own encodings; in fact even the same data source can support more than one,
e.g., Debezium uses a slightly different encoding for each source DB, and
that's just using the default configuration.

To support this, we need to implement configurable deserializers that can
change their behavior based on a user-defined profile.  This should be
straightforward for the JIT deserialization framework that supports
configurable deserializers by design.  For statically compiled rust,
this is harder.  Unfortunately, serde's `Deserialize` trait doesn't allow
configurable deserialization.  The only support `serde` provides for passing
state to deserializers is the `DeserializeSeed` trait, but there is no support
for automatically threading it through nested types or to auto-derive
implementations using it.

* We define `trait DeserializeWithContext`, which has the same signature as
  `Deserialize` with the extra context argument.

* `deserialize_without_context` macro is used to implement `DeserializeWithContext`
  for types that implement `serde::Deserialize` and don't support configurable
  deserialization.

* We implement `DeserializeWithContext` for vectors, tuples, and options
  by passing the context to each element.

* The `deserialize_struct` macro is used to implement `DeserializeWithContext`
  for structs.

* The `deserialize_table_record` macro implements `DeserializeWithContext`
  for SQL table records.  It's similar to `deserialize_struct`, but additionally
  handles case-(in)sensitive SQL columns.

* The `SqlDeserializerConfig` type captures deserializer configuration
  for parsing SQL records.  Initially, it captures `DATE`, `TIME`, and
  `TIMESTAMP` formats sufficient to parse Debezium MySQL JSON.

This solution is not perfect.  In particular, we are no longer able to
use `serde_derive` to derive deserialization logic for all types.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>