Fix parsing CSV data with Windows-style newlines.
The CSV parser library is supposed to recognize `\r`, `\n`, and `\r\n`
as line terminators, but it does it in a strange way with the `\r\n`
being split in the middle.  This caused the parser to recognize the
final `\n` in a block of CSV records as an invalid record.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>