Adjustments for Debezium decimal parsing.
Debezium uses a very unfortunate default representation of decimal
numbers that only encodes the digits, leaving it up to the consumer to
infer the position of the decimal dot.  This cannot be done in a robust
and efficient way.  The most obvious approach is to infer the scale of
the number from the SQL table definitions that we have, but thsese
definitions are not guaranteed to precisely match the source DB, esp
when it comes to decimal precisions.  An alternative is to infer it from
the table schema that Debezium attaches to each record by default.  But
including the schema in each message is exceedingly expensive, so we
should allow (and recommend) the users to disable it, so that's also
not a good solution.

When the schema registry is available, a reference to the schema
registry can be used instead of inlining the schema in each record.  But
that's also problematic for multiple reasons.  First, the schema registy
may not always be configured.  Second, some schema registry
implementations don't support JSON schmas, e.g., Redpanda only supports
Avro and Protobuf.  Overall, while we do want to add support for schema
registies going forward, I strongly prefer to be able to parse JSON
without relying on one.

In summary, I don't see a clean solution at the moment.  As a temporary
solution that doesn't require any extra work on our part, I suggest we
require the user to configure their Debezium connectors to use string
encoding of decimals (which we already support!).  All it takes is
adding one line to the connector config:

```
"decimal.handling.mode": "string",
```

This commit

- Adds the above config option to our Debezium demo.
- Adds decimals to our deserialization tests.
- Clarifies in the docs that we support and prefer string representation
  for decimals in JSON.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>