Add tutorial as the second demo.
I created a new demo based on the "Feldera: The Basics" tutorial.  The
idea is that people who are too lazy to go through all the steps in the
tutorial can just play with the resulting pipeline.

- Added missing Python API to create URL input connectors.
- Added `project_demo06-SupplyChainTutorial` demo, which creates and
  configures the pipeline described in the tutorial using API with one
  differences: it doesn't set Kafka broker address in the connectors, so that the
    pipeline picks them up from the `REDPANDA_BROKERS` env variable.
    This way the demo works both in docker and locally.
- Modified `docker-compose.yml` to create the new demo, along with the
  SecOps demo in the demo profile.
  - In the process, added a dependency on the redpanda container to the
    demo, so it doesn't start until redpanda is up and running, and
    reduced the demo startup timeout from 5 to 1 seconds.  Perhaps even
    that is unnecessary.
- Updated the tutorial to point impatient users to the pre-made demo.

One issue with this is that we now compile two programs when running
with `--profile demo`, so it takes longer for the demo pipeline to become
available.  We could use a separate profile for the tutorial, but that's
going to become confusing and unwieldy.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>