`Catalog::register_input_map()`
Added a method to register input streams created using `add_input_map`
in the catalog.  Such a stream represents a table with a unique or
primary key and upsert modification semantics.  Entries are deleted by
key and updated or modified by value (inserting a value whose key
already exists in the collection removes the existing value associated
with this key).

In addition to the input stream, the caller must provide a function to
extract keys from values.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>