Custom Kafka Connect docker image.
In preparation for adding support for Snowflake output connector, we
replace the pre-built Kafka Connect image from Debezium with a custom
one that adds Snowflake connector classes to the image.  There are a
couple of Kafka Connect base images we could build upon, including one
from Confluent, but that one seems to assume Apache Kafka and may not
work well with Redpanda (first thing it did when I tried it is
complain about not having Zookeeper around :) ).

In addition, we want the image to contain Debezium connectors, so
building on the Debezium image from RedHat seemed like a good choice.

This commit:
1. Adds a Kafka Connect container build step to Earthfile
2. Adds identical instructions for building the connector to Dockerfile
3. Sets up github CI to build the container along with other Feldera
   containers.

While I was at it, I also added a trigger to the `containers.yml` to
allow triggering the container build workflow manually.  This is useful
if we want to build the containers without waiting for the cronjob.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>