pipeline-manager: do not shutdown pipelines automatically on failure
Several state transitions in the pipeline automata shutdown the pipeline
the moment it encounters a failure. This behavior is undesirable in
production settings where pipelines can fail for myriad reasons, and
it's good to have them stick around to be diagnosed.

Pipelines can anyhow only be restarted via the /start action after an
explicit shutdown by the user.

Signed-off-by: Lalith Suresh <lalith@feldera.com>