Use a separate crate for pipeline types that are part of its API (#875)
Currently, the pipeline-manager depends on the adapters crate, but only
to access a few types that are part of the pipeline public interface
(e.g., types to configure pipelines and their connectors, as well as
error types). However, the adapters crate contains a whole lot more of
code and dependencies, all of which are brought when we perform
pipeline-manager builds.

This PR refactors out the public API types in the adapters crate into a
separate pipeline-types crate. This allows the pipeline-manager to only
depend on pipeline-types and not adapters. On my M2 macbook, I can now
do a clean release build for the pipeline manager in a minute compared
to more than 3 minutes from before.

A consequence of this PR is that the pipeline-manager and adapters
crates can now be built in parallel, which should also speed up builds
in the CI.

Signed-off-by: Lalith Suresh <lalith@feldera.com>