Apply arbitrary filters to keys, values in traces.
We generalize the API to GC keys and values in a trace based on arbitrary
predicates.

Motivation
==========

In order to evaluate windowed queries on time series data using bounded memory
we need a mechanism to garbage collect unused keys and values.  Intuitively,
unused values are values whose timestamp is outside of any window whose
contents can change and hence cannot affect the output of any
incremental query.

Previous design
===============

Our previous design for this assumed that values were ordered by time
and the GC API allowed the client to set monotonically increasing bounds
on values, so that all values < bound could be GC'd.  The actual GC was
performed lazily, whem merging batches inside the trace.

The problem with this design is that it requires arranging data so that
it's sorted by time, which (1) can be costly, (2) complicates the
compiler, (3) may be hard or impossible for collections with multiple
timestamps.

New design
==========

In this commit we generalize the GC API to support arbitrary
user-defined predicates (a predicate is a closure that takes a value and
returns `true` if the value is to be retained and `false` otherwise).

GC is still performed lazily at merge time.

The new API is more flexible but potentially less efficient: although the
complexity is still O(1) per record, paid at merge time, having to apply
the check to each value can be expensive compared to simply copying a
slice in memory.  One option is to keep both APIs around, but I don't do
this to keep things simple.

TODO:
- Expose the API via DBSP operators (it is currently only available on
  traces)
- Implement the API for persistent traces (it's currently a noop)

This is part of work on https://github.com/feldera/feldera/issues/873

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>