Don't export the output of Z1Trace.
Strict operators (i.e., operators on the backedge of a feedback loop)
have a feature that allows the parent circuit to consume the last output
of the operator after its clock epoch ends.  This is useful to make
the result of a fixed point computation availabel to the parent.  This
behavior used to be mandatory for all strict operators.  For the `Z1Trace`
operator, when configured to integrate inputs across multiple clock
epochs this meant that it had to manufacture a bogus empty trace just
to keep the API happy.  It turns out that empty traces are expensive
when using RocksDB for persistence.

The fix is to make the export feature optional and not create an export
stream for operators like `Z1Trace`, since these are not used anyway.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>