Remove return type in `push_merge`.
`MergeBuilder::push_merge` and `MergeBuilder::push_merge_retain_keys`
used to return a `usize` value, which was equal to the number of keys in
the builder after the merge, but was misinterpreted in one callsite as
the amount of fuel spent by the methods.

We remove the return types, since the client can just call
`MergeBuilder::keys()` to get the number of keys.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>