Apply retainment policies for traces.
This commit implements an API to specify data retainment policies for
keys and/or values in a trace.

Background
==========

Relations that store time series data typically have the property that any new
updates can only affect records with recent timestamps.  Depending on how the
relation is used in queries this might mean that, while records with older
timestamps still exist in the relation, they cannot affect any future incremental
computation and therefore don't need to be stored.

Design
======

We support two mechanisms to specify and eventually discard such unused
records.

The first (already existing) mechanism, exposed via the
`Stream::integrate_trace_with_bound` method, is only applicable when
keys and/or values in the collection are ordered by time.  It allows
_each_ consumer of the trace to specify a lower bound on the keys and
values it is interested in.  The effective bound is the minimum of all
bounds specified by individual consumers.

The second mechanism, implemented in this commit is more general and
allows the caller to specify arbitrary conditions on keys and values in
the trace.  Keys (values) that don't satisfy the condition are
eventually reclaimed by the trace.  This mechanism is applicable to
collections that are not ordered by time.  Hence it doesn't require
rearranging the data in time order.  Furthermore, it is applicable to
collections that contain multiple timestamp column.  Such
multidimensional timestamps only form a lattice, not a total order.

Unlike the first mechanism, this mechanism only allows one global
condition to be applied to the stream.  This bound affects _all_
operators that use the trace of the stream, i.e., call `integrate_trace`
(or `trace` in the root scope) on it. This includes for instance `join`,
`aggregate`, and `distinct`.  All such operators will reference the same
instance of a trace.  Therefore bounds specified by this API must be
based on a global analysis of the entire program.

The two mechanisms described above interact in different ways for keys
and values.  For keys, the lower bound and the retainment condition are
independent and can be active at the same time.  Internally, they are
enforced using different techniques.  Lower bounds are enforced at
essentially zero cost.  The retention condition is more expensive, but
more general.

For values, only one of the two mechanisms can be enabled for any given
stream.  Whenever a retainment condition is specified it supersedes any
lower bounds constraints.

Correctness
===========

Retainment policies set using the new API apply to all consumers of the
trace.  An incorrect policy may reclaim keys that are still needed by
some of the operators, leading to incorrect results.  Computing a
correct retainment policy can be a subtle and error prone task, which is
probably best left to automatic tools like compilers.

In this commit
=============

* `watermark` operator - computes the lower upper bound of all
 timestamps in the stream without assuming that timestamps form
 a total order or that the data is indexed by time

* `integrate_trace_retain_key`, `integrate_trace_retain_value`
 operators can be used to apply a retainment policy to keys
 and/or values in a trace.

* Along the way, generalized the exchange operator not to require
 the output type to implement `Default`.  The user sets the initial
 value of the accumulator explicitly now.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>